Update Rope Movement

+++ Assets/Scripts/Data/BlockData.cs
@@ -5,7 +5,7 @@ public class BlockData
 {
     public int id;
     public string name;
-    public float maxLength;
+    public int maxLength;
     public float curvature;
     public Vector2 positionStartBlock;
     public Vector2 positionEndBlock;
